task 1: make fn createAppointment a pure fn. task2: put fn createAppointment into file utils.js (so it can serve as a building block to update an appointment in use case 4). Both tasks belong to the implementation of 'use case 2: add appointment'. Now I move on with 'use case 3: delete appointment'.
